requirements.txt
# FastAPI 및 웹 서버
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3

# 데이터베이스
sqlalchemy==2.0.42
pymysql==1.1.1
pymongo==4.10.1

# 유틸리티
python-dotenv==1.1.1
email-validator==2.2.0
httpx==0.28.1

# 데이터 처리 및 분석
pandas==2.3.1
numpy==1.26.4
nltk==3.9.1
scipy==1.11.4
gensim==4.3.2

# 시각화
matplotlib==3.10.3
seaborn==0.13.2

# 웹 크롤링
selenium==4.8.2
webdriver-manager==4.0.2

# 테스트
pytest==8.4.1

# 기타 의존성
pydantic==2.10.5
typing-extensions==4.14.1

# Streamlit 및 LangChain
streamlit==1.32.0
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-upstage==0.0.1
langgraph==0.2.0

# 임베딩 및 벡터 검색
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.0.1
transformers==4.30.0
tokenizers==0.13.3

# 기타 ML 라이브러리
scikit-learn==1.3.0
